requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
Pillow==10.4.0
pymupdf==1.24.10
rapidocr-onnxruntime==1.3.24
